Reverting non-functional linter style, adding test
