+++ nextroute/schema/input.go
@@ -2,8 +2,6 @@
 package schema
 
 import (
-	"encoding/json"
-	"errors"
 	"time"
 )
 
@@ -102,34 +100,4 @@ type Stop struct {
 type Location struct {
 	Lon float64 `json:"lon"`
 	Lat float64 `json:"lat"`
-}
-
-// ConvertCustomData converts the custom data into the given type. If the
-// conversion fails, an error is returned.
-func ConvertCustomData[T any](data any) (T, error) {
-	// Marshal the data again in order to unmarshal it into the correct type.
-	var b []byte
-	var err error
-	if rawCustomData, ok := data.(map[string]any); ok {
-		// Typically, the custom data is a map.
-		b, err = json.Marshal(rawCustomData)
-		if err != nil {
-			return *new(T), err
-		}
-	} else if rawCustomData, ok := data.([]any); ok {
-		// Try slice, if not map.
-		b, err = json.Marshal(rawCustomData)
-		if err != nil {
-			return *new(T), err
-		}
-	} else {
-		return *new(T), errors.New("CustomData is not a map or slice")
-	}
-
-	// Unmarshal the custom data into the given custom type.
-	value := new(T)
-	if err := json.Unmarshal(b, value); err != nil {
-		return *new(T), err
-	}
-	return *value, nil
 }
